+++ src/config.js
@@ -1,5 +1,8 @@
-const pluginName = 'html-bundler-webpack-plugin';
-
 module.exports = {
-  pluginName,
+  // plugin name, must be same as node module name
+  pluginName: 'html-bundler-webpack-plugin',
+  // label in terminal output from plugin
+  pluginLabel: 'HTML Bundler Plugin',
+  // label in terminal output from loader
+  loaderLabel: 'HTML Bundler Loader',
 };